requirements.txt
Package            Version
------------------ ---------
certifi            2023.7.22
charset-normalizer 3.3.2
docutils           0.20.1
htag               0.61.0
idna               3.4
Kivy               2.2.1
Kivy-Garden        0.1.5
mutagen            1.47.0
Pillow             10.1.0
pip                23.2
Pygments           2.16.1
requests           2.31.0
setuptools         68.1.2
tornado            6.3.3
urllib3            2.0.7
